Added an authentication check to make sure that users were signed in before being able to access account only features.
