+++ UI/src/pages/MessagingPage.js
@@ -3,7 +3,7 @@ import ChatList from '../components/ChatList';
 import Chat from '../components/Chat';
 import './MessagingPage.css';
 import {Button} from '@mui/material';
-import {addExampleChats} from '../firebaseFunctions/firebaseFunctions';
+import {AuthCheck} from '../components/AuthCheck';
 
 function MessagingPage() {
   const [selectedChat, setSelectedChat] = useState(null);
@@ -12,20 +12,22 @@ function MessagingPage() {
   //   .catch(e => console.log(e));
 
   return (
-    <div className="messaging-page">
-      <div className="chat-list-wrapper">
-        <ChatList onSelectChat={setSelectedChat} />
+    <AuthCheck>
+      <div className="messaging-page">
+        <div className="chat-list-wrapper">
+          <ChatList onSelectChat={setSelectedChat} />
+        </div>
+        {selectedChat ? (
+          <div className="chat-wrapper">
+            <Chat selectedChat={selectedChat} />
+          </div>
+        ) : (
+          <div className="messaging-page-placeholder">
+            <h2>Select a chat to start messaging!</h2>
+          </div>
+        )}
       </div>
-      {selectedChat ? (
-        <div className="chat-wrapper">
-          <Chat selectedChat={selectedChat} />
-        </div>
-      ) : (
-        <div className="messaging-page-placeholder">
-          <h2>Select a chat to start messaging!</h2>
-        </div>
-      )}
-    </div>
+    </AuthCheck>
   );
 }
