+++ UI/src/pages/SellPage.js
@@ -10,6 +10,7 @@ import {
 } from 'firebase/storage';
 import {getResizedImage} from '../utils/imageUtils';
 import {generateUniqueId} from '../utils/generalUtils'; // Import your CSS for styling
+import {AuthCheck} from '../components/AuthCheck';
 
 const SellPage = ({currentUser}) => {
   const [images, setImages] = useState([]);
@@ -112,69 +113,73 @@ const SellPage = ({currentUser}) => {
   };
 
   return (
-    <div className="rent-item">
-      <h1>Create Rental Listing</h1>
-      <form onSubmit={handleSubmit}>
-        <label>
-          Upload Images (up to 5):
-          <input type="file" multiple onChange={handleImageUpload} />
-        </label>
-        <div className="uploaded-images">
-          {imageURLs.map((url, index) => (
-            <img key={index} src={url} alt={`uploaded-${index}`} />
-          ))}
-        </div>
-        <br />
-        <label>
-          Category:
-          <select value={category} onChange={e => setCategory(e.target.value)}>
-            <option value="">Select a category</option>
-            {categories.map(cat => (
-              <option key={cat} value={cat}>
-                {cat}
-              </option>
+    <AuthCheck>
+      <div className="rent-item">
+        <h1>Create Rental Listing</h1>
+        <form onSubmit={handleSubmit}>
+          <label>
+            Upload Images (up to 5):
+            <input type="file" multiple onChange={handleImageUpload} />
+          </label>
+          <div className="uploaded-images">
+            {imageURLs.map((url, index) => (
+              <img key={index} src={url} alt={`uploaded-${index}`} />
             ))}
-          </select>
-        </label>
-        <br />
-        <label>
-          Price:
-          <input
-            type="number"
-            value={price}
-            onChange={e => setPrice(e.target.value)}
-          />
-        </label>
-        <br />
-        <label>
-          Description:
-          <textarea
-            value={description}
-            onChange={e => setDescription(e.target.value)}
-          />
-        </label>
-        <br />
-        <label>
-          Time Limit (hours):
-          <input
-            type="number"
-            value={timeLimit}
-            onChange={e => setTimeLimit(e.target.value)}
-          />
-        </label>
-        <br />
-        <label>
-          List Date:
-          <input
-            type="date"
-            value={listDate}
-            onChange={e => setListDate(e.target.value)}
-          />
-        </label>
-        <br />
-        <button type="submit">Submit Listing</button>
-      </form>
-    </div>
+          </div>
+          <br />
+          <label>
+            Category:
+            <select
+              value={category}
+              onChange={e => setCategory(e.target.value)}>
+              <option value="">Select a category</option>
+              {categories.map(cat => (
+                <option key={cat} value={cat}>
+                  {cat}
+                </option>
+              ))}
+            </select>
+          </label>
+          <br />
+          <label>
+            Price:
+            <input
+              type="number"
+              value={price}
+              onChange={e => setPrice(e.target.value)}
+            />
+          </label>
+          <br />
+          <label>
+            Description:
+            <textarea
+              value={description}
+              onChange={e => setDescription(e.target.value)}
+            />
+          </label>
+          <br />
+          <label>
+            Time Limit (hours):
+            <input
+              type="number"
+              value={timeLimit}
+              onChange={e => setTimeLimit(e.target.value)}
+            />
+          </label>
+          <br />
+          <label>
+            List Date:
+            <input
+              type="date"
+              value={listDate}
+              onChange={e => setListDate(e.target.value)}
+            />
+          </label>
+          <br />
+          <button type="submit">Submit Listing</button>
+        </form>
+      </div>
+    </AuthCheck>
   );
 };
